Revamp utils imports
