+++ cogs/events/tracking.py
@@ -31,8 +31,7 @@ import cachetools
 import discord
 from discord.ext import commands
 
-from utils.bases.cog import BaseCog
-from utils.context import Context
+from utils import BaseCog, Context
 
 if TYPE_CHECKING:
     from bot import FuryBot
